remove all nucleus-core references from ModelEndpoint
